ListContainer 컴포넌트 만들기(리팩토링), App컴포넌트 정리
- ListContainer 컴포넌트 만들고 코드 리팩토링
- App 컴포넌트 정리
