+++ src/ListContainer.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+import { useDispatch, useSelector } from 'react-redux';
+
+import List from './List';
+
+import { deleteTask } from './actions';
+
+export default function App() {
+  const { tasks } = useSelector((state) => ({
+    tasks: state.tasks,
+  }));
+
+  const dispatch = useDispatch();
+
+  function handleClickDeleteTask(id) {
+    dispatch(deleteTask(id));
+  }
+
+  return (
+    <List
+      tasks={tasks}
+      onClickDelete={handleClickDeleteTask}
+    />
+  );
+}
